fix(backend): eliminar implementación duplicada de anomalías que causaba errores de compilación #1924
Existían dos implementaciones paralelas del sistema de anomalías con tipos
redeclarados (AnomalyType, AnomalySeverity, AnomalyDetectionConfig, AnomalyRecord,
AnomalyConfigResponse). Se consolida en la implementación más completa
(OrderAnomalyDetectionService + OrderAnomalyFunction) y se eliminan los archivos
duplicados (OrderAnomalyModels, OrderAnomalyDetector, BusinessOrderAnomaliesFunction).
